requirements.txt
# Core Django and DRF dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication and Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# API Documentation
drf-yasg==1.21.7

# Data Generation
Faker==19.12.0
factory-boy==3.3.0

# Environment Variables
python-decouple==3.8

# Rate Limiting
django-ratelimit==4.1.0

# Filtering
django-filter==23.4

# Data Export
openpyxl==3.1.2
pandas==2.1.3

# Visualization
plotly==5.17.0

# Development and Testing
pytest-django==4.6.0
pytest==7.4.3
coverage==7.3.2

# WSGI Server
gunicorn==21.2.0

# Utilities
pillow==10.1.0
python-dateutil==2.8.2
